#67 refactor: file 별도의 도메인으로 분리
