+++ src/main/java/com/susanghan_guys/server/file/domain/type/FilesType.java
@@ -1,0 +1,6 @@
+package com.susanghan_guys.server.file.domain.type;
+
+public enum FilesType {
+    PLAN,
+    VIDEO
+}
