+++ src/main/java/com/susanghan_guys/server/file/infrastructure/saver/PdfFileSaver.java
@@ -1,12 +1,12 @@
-package com.susanghan_guys.server.work.infrastructure.saver;
+package com.susanghan_guys.server.file.infrastructure.saver;
 
 import com.susanghan_guys.server.work.domain.AdditionalFile;
-import com.susanghan_guys.server.work.domain.PdfFile;
-import com.susanghan_guys.server.work.domain.PdfImage;
+import com.susanghan_guys.server.file.domain.PdfFile;
+import com.susanghan_guys.server.file.domain.PdfImage;
 import com.susanghan_guys.server.work.domain.type.SourceType;
-import com.susanghan_guys.server.work.infrastructure.mapper.PdfFileMapper;
-import com.susanghan_guys.server.work.infrastructure.persistence.PdfFileRepository;
-import com.susanghan_guys.server.work.infrastructure.persistence.PdfImageRepository;
+import com.susanghan_guys.server.file.infrastructure.mapper.PdfFileMapper;
+import com.susanghan_guys.server.file.infrastructure.persistence.PdfFileRepository;
+import com.susanghan_guys.server.file.infrastructure.persistence.PdfImageRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
