+++ src/main/java/com/susanghan_guys/server/personalwork/application/PersonalWorkService.java
@@ -6,12 +6,12 @@ import com.susanghan_guys.server.personalwork.application.port.WorkSummaryPort;
 import com.susanghan_guys.server.personalwork.application.validator.PersonalWorkValidator;
 import com.susanghan_guys.server.personalwork.dto.response.WorkSummaryResponse;
 import com.susanghan_guys.server.user.domain.User;
-import com.susanghan_guys.server.work.domain.PdfFile;
-import com.susanghan_guys.server.work.domain.PdfImage;
+import com.susanghan_guys.server.file.domain.PdfFile;
+import com.susanghan_guys.server.file.domain.PdfImage;
 import com.susanghan_guys.server.work.exception.WorkException;
 import com.susanghan_guys.server.work.exception.code.WorkErrorCode;
-import com.susanghan_guys.server.work.infrastructure.persistence.PdfFileRepository;
-import com.susanghan_guys.server.work.infrastructure.persistence.PdfImageRepository;
+import com.susanghan_guys.server.file.infrastructure.persistence.PdfFileRepository;
+import com.susanghan_guys.server.file.infrastructure.persistence.PdfImageRepository;
 import com.susanghan_guys.server.work.infrastructure.persistence.WorkRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
