+++ src/main/java/com/susanghan_guys/server/work/application/YccWorkService.java
@@ -5,13 +5,13 @@ import com.susanghan_guys.server.global.s3.application.S3Service;
 import com.susanghan_guys.server.global.security.CurrentUserProvider;
 import com.susanghan_guys.server.user.domain.User;
 import com.susanghan_guys.server.work.application.support.WorkHelper;
-import com.susanghan_guys.server.work.domain.PdfFile;
+import com.susanghan_guys.server.file.domain.PdfFile;
 import com.susanghan_guys.server.work.domain.Work;
 import com.susanghan_guys.server.work.domain.type.SourceType;
 import com.susanghan_guys.server.work.dto.request.YccWorkSubmissionRequest;
-import com.susanghan_guys.server.work.infrastructure.mapper.PdfFileMapper;
+import com.susanghan_guys.server.file.infrastructure.mapper.PdfFileMapper;
 import com.susanghan_guys.server.work.infrastructure.mapper.YccWorkMapper;
-import com.susanghan_guys.server.work.infrastructure.persistence.PdfFileRepository;
+import com.susanghan_guys.server.file.infrastructure.persistence.PdfFileRepository;
 import com.susanghan_guys.server.work.infrastructure.persistence.WorkRepository;
 import com.susanghan_guys.server.work.infrastructure.saver.WorkSaver;
 import com.susanghan_guys.server.work.application.validator.YccWorkValidator;
