--- src/main/java/com/susanghan_guys/server/work/domain/type/FilesType.java
@@ -1,6 +1,0 @@
-package com.susanghan_guys.server.work.domain.type;
-
-public enum FilesType {
-    PLAN,
-    VIDEO
-}
