+++ src/main/java/com/susanghan_guys/server/work/presentation/WorkController.java
@@ -2,7 +2,7 @@ package com.susanghan_guys.server.work.presentation;
 
 import com.susanghan_guys.server.global.common.CommonResponse;
 import com.susanghan_guys.server.work.application.DcaWorkService;
-import com.susanghan_guys.server.work.application.PdfFileService;
+import com.susanghan_guys.server.file.application.PdfFileService;
 import com.susanghan_guys.server.work.application.YccWorkService;
 import com.susanghan_guys.server.work.dto.request.DcaWorkSubmissionRequest;
 import com.susanghan_guys.server.work.dto.request.YccWorkSubmissionRequest;
@@ -23,7 +23,6 @@ public class WorkController implements WorkSwagger {
 
     private final DcaWorkService dcaWorkService;
     private final YccWorkService yccWorkService;
-    private final PdfFileService pdfImageService;
 
     @Override
     @PostMapping(value = "/dca", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
@@ -43,20 +42,6 @@ public class WorkController implements WorkSwagger {
             @RequestPart MultipartFile planFile
     ) {
         yccWorkService.submit(request, planFile);
-        return CommonResponse.success(WORK_YCC_SUBMIT_SUCCESS, "OK");
-    }
-
-    // dca - test
-    @PostMapping("/dca/{workId}")
-    public CommonResponse<String> convertDcaPdfToImage(@PathVariable("workId") Long workId) {
-        pdfImageService.convertDcaPdfToImage(workId);
-        return CommonResponse.success(WORK_DCA_SUBMIT_SUCCESS, "OK");
-    }
-
-    // ycc - test
-    @PostMapping("/ycc/{workId}")
-    public CommonResponse<String> convertToImage(@PathVariable("workId") Long workId) {
-        pdfImageService.convertYccPdfToImage(workId);
         return CommonResponse.success(WORK_YCC_SUBMIT_SUCCESS, "OK");
     }
 }
